Update pet classes to use float for age instead of int. Started implementing error handling, will create a method for this.

+++ PetAdoptionCentre/Bird.cs
@@ -1,5 +1,5 @@
 namespace PetAdoptionCentre {
-    public class Bird(string name, int age, string type, bool isAdopted, string health) 
+    public class Bird(string name, float age, string type, bool isAdopted, string health) 
                           :Pet(name, age, type, isAdopted, health) {
 
         // You have to use override when defining abstract implied methods when inheriting
@@ -8,4 +8,4 @@ namespace PetAdoptionCentre {
         }
 
     }
-}   
+}
