+++ PetAdoptionCentre/Dog.cs
@@ -1,5 +1,5 @@
 namespace PetAdoptionCentre {
-    public class Dog(string name, int age, string type, bool isAdopted, string health) 
+    public class Dog(string name, float age, string type, bool isAdopted, string health) 
                           :Pet(name, age, type, isAdopted, health) {
 
         // You have to use override when defining abstract implied methods when inheriting
@@ -8,4 +8,4 @@ namespace PetAdoptionCentre {
         }
 
     }
-}   
+}
